Dev samuel

+++ resources/views/content/medicos/horario.blade.php
@@ -31,7 +31,7 @@
                         @php
                             $horario = $medico->horarios->firstWhere('dia_semana', $dia);
                         @endphp
-                        <tr>
+                        <tr id="fila-{{ $dia }}">
                             <td>{{ $dia }}</td>
                             <td>
                                 <input type="time" name="horarios[{{ $dia }}][hora_inicio]"
@@ -48,12 +48,16 @@
                                     value="{{ $horario->descanso_termino ?? '' }}" class="form-control mt-2">
                             </td>
                             <td>
-                                <input type="text" name="horarios[{{ $dia }}][box_atencion]"
-                                    value="{{ $horario->box_atencion ?? '' }}" class="form-control">
+                                <select name="box" id="box" class="form-control">
+                                    <option value="">Seleccione un box</option>
+                                    @foreach($boxes as $box)
+                                        <option value="{{ $box->nombre }}">{{ $box->nombre }}</option>
+                                    @endforeach
+                                </select>
                             </td>
                             <td>
-                                <input type="checkbox" name="horarios[{{ $dia }}][no_atiende]" value="1"
-                                    {{ $horario->no_atiende ? 'checked' : '' }}>
+                                <input type="checkbox" name="horarios[{{ $dia }}][no_atiende]" value="1" 
+                                {{ $horario && $horario->no_atiende == 1 ? 'checked' : '' }} class="no-atiende-checkbox">
                             </td>
                         </tr>
                     @endforeach
@@ -69,8 +73,9 @@
         + Crear un nuevo bloqueo programado
     </button>
     <br>
-    <div class="row">
-        <table id="bloqueosTable">
+    
+    <div class="card-datatable table-responsive pt-0">
+        <table class="dt-responsive table table" id="bloqueosTable">
             <thead>
                 <tr>
                     <th>Sucursal</th>
@@ -104,7 +109,13 @@
                     <form>
                         <div class="form-group">
                             <label for="sucursal">Sucursal:</label>
-                            <input type="text" class="form-control" name="sucursal" id="sucursal" required>
+                            <select name="sucursal" id="sucursal" class="form-control">
+                                <option value="">Seleccione una sucursal</option>
+                                @foreach($sucursales as $sucursal)
+                                    <option value="{{ $sucursal->nombre }}">{{ $sucursal->nombre }}</option>
+                                @endforeach
+                            </select>
+
                         </div>
                         <div class="form-group">
                             <label for="fecha">Fecha:</label>
@@ -118,6 +129,15 @@
                             <label for="hora_termino">Hora Término:</label>
                             <input type="time" class="form-control" name="hora_termino" id="hora_termino" required>
                         </div>
+                        <div class="form-group">
+                            <label for="recurso">Box</label>
+                            <select name="recurso" id="recurso" class="form-control">
+                                <option value="">Seleccione un box</option>
+                                @foreach($boxes as $box)
+                                    <option value="{{ $box->nombre }}">{{ $box->nombre }}</option>
+                                @endforeach
+                            </select>
+                        </div>
                         <button type="button" id="bloqueoForm" class="btn btn-primary">Guardar</button>
                     </form>
                 </div>
@@ -127,25 +147,28 @@
 
 @endsection
 
+<script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
+
+
 
 <script>
     $(document).ready(function() {
         const medicoId = {{ $medico->id }};
-        alert();
         cargarBloqueos();
 
         // Agregar nuevo bloqueo
         $('#bloqueoForm').on('click', function(e) { // Cambié 'onClick' por 'click'
             e.preventDefault();
             console.log('BloqueoForm button clicked'); // Para verificar que el evento se activa
+            window.userId = {{ auth()->user()->id }};
 
             const formData = {
                 sucursal: $('#sucursal').val(),
                 fecha: $('#fecha').val(),
                 hora_inicio: $('#hora_inicio').val(),
                 hora_termino: $('#hora_termino').val(),
-                recurso: 1,
-                creado_por: 'Usuario Actual', // Cambiar según tu lógica
+                recurso: $('#recurso').val(),
+                creado_por: window.userId, // Cambiar según tu lógica
             };
 
             console.log('Form data:', formData);
@@ -159,7 +182,6 @@
                     'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content'),
                 },
                 success: function(response) {
-                    alert('Bloqueo guardado correctamente.');
                     $('#bloqueoModal').modal('hide'); // Cerrar el modal
                     // cargarBloqueos(); // Actualizar la tabla
                 },
@@ -186,7 +208,7 @@
                             <td>${bloqueo.fecha}</td>
                             <td>${bloqueo.hora_inicio}</td>
                             <td>${bloqueo.hora_termino}</td>
-                            <td>${bloqueo.creado_por}</td>
+                            <td>${bloqueo.name}</td>
                             <td>${bloqueo.recurso}</td>
                             <td>
                                 <button class="btn btn-danger btn-sm delete-btn" data-id="${bloqueo.id}">
@@ -212,21 +234,66 @@
 
         // Eliminar bloqueo
         function eliminarBloqueo(id) {
-            $.ajax({
-                url: `/bloqueos/${id}`,
-                method: 'DELETE',
-                headers: {
-                    'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content'),
-                },
-                success: function() {
-                    alert('Bloqueo eliminado correctamente.');
-                    cargarBloqueos();
-                },
-                error: function(xhr) {
-                    console.error(xhr.responseText);
-                    alert('Error al eliminar el bloqueo.');
-                },
+            Swal.fire({
+                title: "¿Estás seguro?",
+                text: "¡No podrás revertir esta acción!",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Sí, eliminarlo",
+                cancelButtonText: "Cancelar"
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    $.ajax({
+                        url: `/bloqueos/${id}`,
+                        method: 'DELETE',
+                        headers: {
+                            'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content'),
+                        },
+                        success: function() {
+                            Swal.fire({
+                                title: "Eliminado!",
+                                text: "El bloqueo ha sido eliminado correctamente.",
+                                icon: "success"
+                            });
+                            cargarBloqueos();
+                        },
+                        error: function(xhr) {
+                            console.error(xhr.responseText);
+                            Swal.fire({
+                                title: "Error",
+                                text: "Hubo un problema al eliminar el bloqueo.",
+                                icon: "error"
+                            });
+                        },
+                    });
+                }
             });
         }
+
+    });
+</script>
+
+<script>
+    $(document).ready(function() {
+        // Función para deshabilitar/activar la fila completa
+        $('.no-atiende-checkbox').change(function() {
+            var fila = $(this).closest('tr');  // Obtiene la fila correspondiente
+            if ($(this).is(':checked')) {
+                // Si el checkbox está marcado, deshabilitar todos los inputs y selects de la fila, excepto el checkbox
+                fila.find('input[type="time"], select').prop('disabled', true);
+            } else {
+                // Si el checkbox no está marcado, habilitar todos los inputs y selects de la fila
+                fila.find('input[type="time"], select').prop('disabled', false);
+            }
+        });
+
+        // Inicialización para marcar las filas ya deshabilitadas según la condición
+        $('.no-atiende-checkbox').each(function() {
+            if ($(this).is(':checked')) {
+                $(this).closest('tr').find('input[type="time"], select').prop('disabled', true);
+            }
+        });
     });
 </script>
